Factor out ImageData.resize into ImageResizer (FINERACT-1201)
