+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/documentmanagement/data/ImageData.java
@@ -18,20 +18,12 @@
  */
 package org.apache.fineract.infrastructure.documentmanagement.data;
 
-import java.awt.Color;
-import java.awt.Graphics2D;
-import java.awt.RenderingHints;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.IOException;
+import java.io.FileNotFoundException;
 import java.io.InputStream;
-import javax.imageio.ImageIO;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.fineract.infrastructure.documentmanagement.contentrepository.ContentRepositoryUtils;
-import org.apache.fineract.infrastructure.documentmanagement.contentrepository.ContentRepositoryUtils.ImageFileExtension;
 import org.apache.fineract.infrastructure.documentmanagement.domain.StorageType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,51 +44,6 @@ public class ImageData {
         this.location = location;
         this.storageType = storageType;
         this.entityDisplayName = entityDisplayName;
-    }
-
-    public ImageData resize(ImageData image, Integer maxWidth, Integer maxHeight) {
-        if (maxWidth == null && maxHeight != null) {
-            return image;
-        }
-        try (InputStream is = image.getInputStream()) {
-            InputStream resizedIS = resizeImage(image.getFileExtension(), is, maxWidth != null ? maxWidth : Integer.MAX_VALUE,
-                    maxHeight != null ? maxHeight : Integer.MAX_VALUE);
-
-            ImageData resizedImage = new ImageData(image.location(), image.storageType(), image.getEntityDisplayName());
-            resizedImage.updateContent(resizedIS);
-            return resizedImage;
-        } catch (IOException e) {
-            LOG.warn("resize() failed, returning original image: {}", e.getMessage(), e);
-            return image;
-        }
-    }
-
-    private InputStream resizeImage(ImageFileExtension fileExtension, InputStream in, int maxWidth, int maxHeight) throws IOException {
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-
-        BufferedImage src = ImageIO.read(in);
-        if (src.getWidth() <= maxWidth && src.getHeight() <= maxHeight) {
-            in.reset();
-            return in;
-        }
-        float widthRatio = (float) src.getWidth() / maxWidth;
-        float heightRatio = (float) src.getHeight() / maxHeight;
-        float scaleRatio = widthRatio > heightRatio ? widthRatio : heightRatio;
-
-        // TODO(lindahl): Improve compressed image quality (perhaps quality ratio)
-
-        int newWidth = (int) (src.getWidth() / scaleRatio);
-        int newHeight = (int) (src.getHeight() / scaleRatio);
-        int colorModel = fileExtension == ContentRepositoryUtils.ImageFileExtension.JPEG ? BufferedImage.TYPE_INT_RGB
-                : BufferedImage.TYPE_INT_ARGB;
-        BufferedImage target = new BufferedImage(newWidth, newHeight, colorModel);
-        Graphics2D g = target.createGraphics();
-        g.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR);
-        g.drawImage(src, 0, 0, newWidth, newHeight, Color.BLACK, null);
-        g.dispose();
-        ImageIO.write(target, fileExtension != null ? fileExtension.getValueWithoutDot() : "jpeg", os);
-
-        return new ByteArrayInputStream(os.toByteArray());
     }
 
     private void setImageFileExtension(String filename) {
